Refactor signup slice
